# packages/contracts/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "ContractMetadata": {
        "additionalProperties": false,
        "description": "Mode, fraîcheur et version d'une réponse métier.",
        "properties": {
          "appVersion": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Appversion",
            "type": "string"
          },
          "asOf": {
            "format": "date-time",
            "title": "Asof",
            "type": "string"
          },
          "computedAt": {
            "format": "date-time",
            "title": "Computedat",
            "type": "string"
          },
          "dataMode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "freshness": {
            "$ref": "#/components/schemas/FreshnessStatus"
          }
        },
        "required": [
          "dataMode",
          "freshness",
          "asOf",
          "computedAt",
          "appVersion"
        ],
        "title": "ContractMetadata",
        "type": "object"
      },
...
      "DataMode": {
        "description": "Origine isolée des données applicatives.",
        "enum": [
          "mock",
          "real"
        ],
        "title": "DataMode",
        "type": "string"
      },
...
      "Event": {
        "additionalProperties": false,
        "description": "Événement canonique sans dépendance à un fournisseur.",
        "properties": {
          "bestOf": {
            "maximum": 9.0,
            "minimum": 1.0,
            "title": "Bestof",
            "type": "integer"
          },
          "competition": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Competition",
            "type": "string"
          },
          "eventId": {
            "format": "uuid",
            "title": "Eventid",
            "type": "string"
          },
          "gameTitle": {
            "$ref": "#/components/schemas/GameTitle"
          },
          "observedAt": {
            "format": "date-time",
            "title": "Observedat",
            "type": "string"
          },
          "startsAt": {
            "format": "date-time",
            "title": "Startsat",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          },
          "teamA": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Teama",
            "type": "string"
          },
          "teamAId": {
            "format": "uuid",
            "title": "Teamaid",
            "type": "string"
          },
          "teamB": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Teamb",
            "type": "string"
          },
          "teamBId": {
            "format": "uuid",
            "title": "Teambid",
            "type": "string"
          }
        },
        "required": [
          "eventId",
          "gameTitle",
          "competition",
          "teamAId",
          "teamA",
          "teamBId",
          "teamB",
          "startsAt",
          "bestOf",
          "status",
          "observedAt"
        ],
        "title": "Event",
        "type": "object"
      },
      "EventStatus": {
        "description": "État annoncé d'un événement.",
        "enum": [
          "scheduled",
          "live",
          "finished",
          "cancelled"
        ],
        "title": "EventStatus",
        "type": "string"
      },
...
      "FreshnessStatus": {
        "description": "Fraîcheur normative d'une source ou d'un résultat.",
        "enum": [
          "fresh",
          "stale",
          "degraded",
          "failed",
          "quarantined"
        ],
        "title": "FreshnessStatus",
        "type": "string"
      },
      "GameTitle": {
        "description": "Jeux activés dans le périmètre courant.",
        "enum": [
          "lol"
        ],
        "title": "GameTitle",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageInfo": {
        "additionalProperties": false,
        "description": "Fenêtre de pagination stable pour toute collection.",
        "properties": {
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "offset",
          "limit",
          "total"
        ],
        "title": "PageInfo",
        "type": "object"
      },
...
      "PageResponse_Event_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ContractMetadata"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "data",
          "page",
          "meta"
        ],
        "title": "PageResponse[Event]",
        "type": "object"
      },
...
      "PageResponse_StakePublicEvent_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/StakePublicEvent"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ContractMetadata"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "data",
          "page",
          "meta"
        ],
        "title": "PageResponse[StakePublicEvent]",
        "type": "object"
      },
...
      "ProviderEvent": {
        "additionalProperties": false,
        "description": "Projection minimale d'un événement dans le vocabulaire du fournisseur.",
        "properties": {
          "bestOf": {
            "anyOf": [
              {
                "maximum": 9.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bestof"
          },
          "collectedAt": {
            "format": "date-time",
            "title": "Collectedat",
            "type": "string"
          },
          "competition": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Competition",
            "type": "string"
          },
          "gameTitle": {
            "$ref": "#/components/schemas/GameTitle"
          },
          "participants": {
            "items": {
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            },
            "minItems": 2,
            "title": "Participants",
            "type": "array"
          },
          "providerEventId": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Providereventid",
            "type": "string"
          },
          "sourceReference": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Sourcereference",
            "type": "string"
          },
          "startsAt": {
            "format": "date-time",
            "title": "Startsat",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          }
        },
        "required": [
          "providerEventId",
          "gameTitle",
          "competition",
          "participants",
          "startsAt",
          "status",
          "collectedAt",
          "sourceReference"
        ],
        "title": "ProviderEvent",
        "type": "object"
      },
...
      "ScrapedMarket": {
        "additionalProperties": false,
        "properties": {
          "capturedAt": {
            "format": "date-time",
            "title": "Capturedat",
            "type": "string"
          },
          "expanded": {
            "title": "Expanded",
            "type": "boolean"
          },
          "label": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/ScrapedOutcome"
            },
            "maxItems": 512,
            "title": "Outcomes",
            "type": "array"
          },
          "rawText": {
            "maxLength": 20000,
            "title": "Rawtext",
            "type": "string"
          },
          "tab": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Tab",
            "type": "string"
          }
        },
        "required": [
          "label",
          "tab",
          "capturedAt",
          "expanded",
          "rawText",
          "outcomes"
        ],
        "title": "ScrapedMarket",
        "type": "object"
      },
      "ScrapedOutcome": {
        "additionalProperties": false,
        "properties": {
          "decimalOdds": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decimalodds"
          },
          "displayedLabel": {
            "maxLength": 512,
            "title": "Displayedlabel",
            "type": "string"
          },
          "label": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "oddsText": {
            "maxLength": 64,
            "title": "Oddstext",
            "type": "string"
          },
          "status": {
            "enum": [
              "open",
              "suspended",
              "unavailable"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "label",
          "displayedLabel",
          "oddsText",
          "decimalOdds",
          "status"
        ],
        "title": "ScrapedOutcome",
        "type": "object"
      },
...
      "StakeEventCapture": {
        "additionalProperties": false,
        "properties": {
          "event": {
            "$ref": "#/components/schemas/ProviderEvent"
          },
          "expectedTabs": {
            "items": {
              "type": "string"
            },
            "maxItems": 16,
            "title": "Expectedtabs",
            "type": "array"
          },
          "markets": {
            "items": {
              "$ref": "#/components/schemas/ScrapedMarket"
            },
            "maxItems": 512,
            "title": "Markets",
            "type": "array"
          },
          "observedAt": {
            "format": "date-time",
            "title": "Observedat",
            "type": "string"
          },
          "sourceUrl": {
            "maxLength": 512,
            "title": "Sourceurl",
            "type": "string"
          },
          "visitedTabs": {
            "items": {
              "type": "string"
            },
            "maxItems": 16,
            "title": "Visitedtabs",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "event",
          "sourceUrl",
          "observedAt",
          "expectedTabs",
          "visitedTabs",
          "markets",
          "warnings"
        ],
        "title": "StakeEventCapture",
        "type": "object"
      },
      "StakePublicEvent": {
        "additionalProperties": false,
        "properties": {
          "ageSeconds": {
            "minimum": 0.0,
            "title": "Ageseconds",
            "type": "integer"
          },
          "capture": {
            "$ref": "#/components/schemas/StakeEventCapture"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "freshness": {
            "$ref": "#/components/schemas/FreshnessStatus"
          }
        },
        "required": [
          "capture",
          "ageSeconds",
          "expiresAt",
          "freshness"
        ],
        "title": "StakePublicEvent",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Metiquo API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "operationId": "list_events_api_v1_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "competition",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Competition"
            }
          },
          {
            "in": "query",
            "name": "team",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Team"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EventStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "startsFrom",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Startsfrom"
            }
          },
          {
            "in": "query",
            "name": "startsTo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Startsto"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_Event_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Events",
        "tags": [
          "reads"
        ]
      }
    },
...
    "/api/v1/odds/stake/events": {
      "get": {
        "operationId": "events_api_v1_odds_stake_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "startsFrom",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Startsfrom"
            }
          },
          {
            "in": "query",
            "name": "startsTo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Startsto"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_StakePublicEvent_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events",
        "tags": [
          "stake-scraping"
        ]
      }
    },
...
  }
}